[#21] 버튼 상태 다른 파일로 분리
